+++ CHANGELOG.md
@@ -7,4 +7,6 @@
   - search data
   - create database if not exists
   - create class object for Base, no need to define tablename
+  - no need to import Base from dbjson
+  - add multiple data